fixed getFontSize command in Scripter for decimal points
git-svn-id: svn://scribus.net/trunk/Scribus@20588 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/scriptplugin/cmdtext.cpp
@@ -69,7 +69,7 @@ PyObject *scribus_getfontsize(PyObject* /* self */, PyObject* args)
 		return NULL;
 	}
 	else
-		return PyFloat_FromDouble(static_cast<long>(it->currentCharStyle().fontSize() / 10.0));
+		return PyFloat_FromDouble(static_cast<double>(it->currentCharStyle().fontSize() / 10.0));
 }
 
 PyObject *scribus_getfont(PyObject* /* self */, PyObject* args)
